Add files via upload
This project demonstrates how to perform face detection using a pretrained model in OpenCV. We use the cv2.CascadeClassifier function, which takes the path to an XML file that contains the pretrained model. These XML files are typically trained using Haar features or LBP (Local Binary Patterns) and are available in OpenCV’s data repository.